+++ mayan/apps/fundraisers/dependencies.py
@@ -1,5 +1,3 @@
 from mayan.apps.dependencies.classes import PythonDependency
 
 PythonDependency(module=__name__, name='Markdown', version_string='==3.9')
-
-PythonDependency(module=__name__, name='nh3', version_string='==0.3.0')