+++ mayan/apps/messaging/model_mixins.py
@@ -1,6 +1,6 @@
-from bleach import Cleaner
-from bleach.linkifier import LinkifyFilter
+import nh3
 
+from django.utils.html import urlize
 from django.utils.translation import gettext_lazy as _
 
 from mayan.apps.templating.template_backends import Template
@@ -16,11 +16,15 @@ class MessageBusinessLogicMixin:
     get_label.short_description = _(message='Label')
 
     def get_clean_body(self):
-        cleaner = Cleaner(
-            filters=[LinkifyFilter]
+        output_linkified = urlize(
+            autoescape=True, nofollow=False, text=self.body
         )
 
-        return cleaner.clean(text=self.body)
+        output_cleaned = nh3.clean(
+            html=output_linkified, link_rel='nofollow noopener noreferrer'
+        )
+
+        return output_cleaned
 
     def get_rendered_body(self):
         clean_body = self.get_clean_body()